requirements.txt
django==1.11
pymysql
djangorestframework==3.9.4
